+++ include/managers/views/crowpanel_audio_view.h
@@ -1,0 +1,8 @@
+#ifndef CROWPANEL_AUDIO_VIEW_H
+#define CROWPANEL_AUDIO_VIEW_H
+
+#include "managers/display_manager.h"
+
+extern View crowpanel_audio_view;
+
+#endif